Have CapabilitiesEventHandler use correct bus
